Move channel info to new component
