+++ src/types/importCandidate.types.ts
@@ -12,7 +12,7 @@ import { MapPublicationInstance } from './publication_types/otherRegistration.ty
 import { PresentationPublicationInstance } from './publication_types/presentationRegistration.types';
 import { ReportPublicationInstance } from './publication_types/reportRegistration.types';
 import { ResearchDataPublicationInstance } from './publication_types/researchDataRegistration.types';
-import { Journal, Publisher, Registration, RegistrationAggregations, Series } from './registration.types';
+import { Journal, Publisher, Registration, RegistrationAggregations } from './registration.types';
 
 export type ImportCandidateStatus = 'IMPORTED' | 'NOT_IMPORTED' | 'NOT_APPLICABLE';
 
@@ -51,7 +51,6 @@ export interface ImportCandidateSummary {
   organizations: (Pick<Organization, 'type' | 'id' | 'labels'> | UnconfirmedOrganization)[];
   publisher?: Partial<Publisher>;
   journal?: Partial<Journal>;
-  series?: Partial<Series>;
   publicationInstance: PublicationInstance;
   contributors: Contributor[];
   printIssn?: string;
